+++ Properties/AssemblyInfo.cs
@@ -31,6 +31,6 @@ using System.Runtime.Versioning;
 //      Numero di build
 //      Revisione
 //
-[assembly: AssemblyVersion("0.9.2.3")]
-[assembly: AssemblyFileVersion("0.9.2.3")]
-[assembly: AssemblyInformationalVersion("0.9.2.3")]
+[assembly: AssemblyVersion("0.9.2.4")]
+[assembly: AssemblyFileVersion("0.9.2.4")]
+[assembly: AssemblyInformationalVersion("0.9.2.4")]